+++ src/main/java/com/example/demo/service/MemberService.java
@@ -2,6 +2,10 @@ package com.example.demo.service;
 
 import java.util.List;
 
+import com.example.demo.exception.EmailExistenceException;
+import com.example.demo.exception.ExistArticleException;
+import com.example.demo.exception.MemberNotFoundException;
+import com.example.demo.repository.ArticleRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,13 +20,16 @@ import com.example.demo.repository.MemberRepository;
 public class MemberService {
 
     private final MemberRepository memberRepository;
+    private final ArticleRepository articleRepository;
 
-    public MemberService(MemberRepository memberRepository) {
+    public MemberService(MemberRepository memberRepository, ArticleRepository articleRepository) {
         this.memberRepository = memberRepository;
+        this.articleRepository = articleRepository;
     }
 
     public MemberResponse getById(Long id) {
-        Member member = memberRepository.findById(id);
+        Member member = memberRepository.findById(id)
+                .orElseThrow(()-> new MemberNotFoundException("사용자 조회 실패"));
         return MemberResponse.from(member);
     }
 
@@ -43,12 +50,20 @@ public class MemberService {
 
     @Transactional
     public void delete(Long id) {
+        if(articleRepository.existsByAuthorId(id)){
+            throw new ExistArticleException("사용자가 작성한 게시물이 존재합니다.");
+        }
         memberRepository.deleteById(id);
     }
 
     @Transactional
     public MemberResponse update(Long id, MemberUpdateRequest request) {
-        Member member = memberRepository.findById(id);
+        Member member = memberRepository.findById(id)
+                .orElseThrow(()-> new MemberNotFoundException("사용자 조회 실패"));
+
+        if (!member.getEmail().equals(request.email()) && memberRepository.existsByEmail(request.email())) {
+            throw new EmailExistenceException("이미 존재하는 이메일입니다: " + request.email());
+        }
         member.update(request.name(), request.email());
         memberRepository.update(member);
         return MemberResponse.from(member);
